+++ app/models/invoices/invoice.rb
@@ -37,15 +37,5 @@
 module Invoices
   class Invoice < ::Invoice
     ::Invoice.register_subtype self
-
-    def suggested_invoice_parts
-      deposits = Invoices::Deposit.of(booking).kept
-      I18n.with_locale(booking.locale) do
-        super + [InvoiceParts::Add.new(
-          apply: new_record?, label: Invoices::Deposit.model_name.human,
-          amount: - deposits.sum(&:amount_paid)
-        )]
-      end
-    end
   end
 end